+++ app/src/main/java/com/example/rito/sitaca/PengumumanActivity.java
@@ -134,7 +134,15 @@ public class PengumumanActivity extends ActionBarActivity {
                         .setPositiveButton("Iya", new DialogInterface.OnClickListener() {
 
                             public void onClick(DialogInterface dialog, int whichButton) {
-                                fragment.delete();
+                                if (new Connection().checkConnection(getApplicationContext()))
+                                    fragment.delete();
+                                else {
+                                    Toast.makeText(
+                                            fragment.rootView.getContext(),
+                                            "Kesalahan : Anda tidak tersambung ke internet.",
+                                            Toast.LENGTH_SHORT
+                                    ).show();
+                                }
                             }
                         })
                         .setNegativeButton("Tidak", null).show();
@@ -326,10 +334,12 @@ public class PengumumanActivity extends ActionBarActivity {
                                     o.getString("waktu")
                             );
                             listPengumuman.add(pengumuman);
+
                         } catch (JSONException e) {
                             e.printStackTrace();
                         }
                     }
+
 
                     Collections.reverse(listPengumuman);
                     adapterPengumuman = new PengumumanListAdapter(listPengumuman);
